+++ sdk/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.1.0 - 2023-MM-DD
 
+### Added
+
+- `Output::kind_str()` method;
+
 ### Changed
 
 - `StrongholdAdapterBuilder` updated to be slightly more ergonomic;